Moved unit-level util modules to util.unit package; related code and test updates; updated trigger for UI CI workflow
