+++ .github/workflows/tkltest_ui_ci.yml
@@ -2,21 +2,21 @@ name: TackleTest-UI CI
 
 on:
   push:
-    paths:
-      - 'tkltest/util/**'
-      - 'tkltest/tkltest.py'
-      - 'tkltest/tkltest_ui.py'
-      - 'test/ui/**'
-      - '.github/workflows/tkltest_ui_ci.yml'
-      - 'setup.*'
+    paths-ignore:
+      - 'tkltest/generate/unit/**'
+      - 'tkltest/execute/unit/**'
+      - 'tkltest/util/unit/**'
+      - 'tkltest/tkltest_unit.py'
+      - 'doc/**'
+      - '**.md'
   pull_request:
-    paths:
-      - 'tkltest/util/**'
-      - 'tkltest/tkltest.py'
-      - 'tkltest/tkltest_ui.py'
-      - 'test/ui/**'
-      - '.github/workflows/tkltest_ui_ci.yml'
-      - 'setup.*'
+    paths-ignore:
+      - 'tkltest/generate/unit/**'
+      - 'tkltest/execute/unit/**'
+      - 'tkltest/util/unit/**'
+      - 'tkltest/tkltest_unit.py'
+      - 'doc/**'
+      - '**.md'
 
 jobs:
   build:
