+++ tkltest/generate/unit/augment.py
@@ -18,7 +18,8 @@ import shutil
 import subprocess
 import sys
 
-from tkltest.util import constants, coverage_util
+from tkltest.util import constants
+from tkltest.util.unit import coverage_util
 from tkltest.util.logging_util import tkltest_status
