+++ tkltest/util/config_util.py
@@ -26,7 +26,8 @@ import json
 from . import constants, config_options
 from .logging_util import tkltest_status
 from .constants import *
-from tkltest.util import command_util, dir_util
+from tkltest.util import command_util
+from tkltest.util.unit import dir_util
 
 
 def load_config(test_level='unit', args=None, config_file=None):
